+++ unlock-js/scripts/compressAbi.js
@@ -16,7 +16,7 @@ const v1 = require('@unlock-protocol/unlock-abi-1')
 const v2 = require('@unlock-protocol/unlock-abi-2')
 const v3 = require('@unlock-protocol/unlock-abi-3')
 const v4 = require('@unlock-protocol/unlock-abi-4')
-const v12 = require('@unlock-protocol/unlock-abi-5')
+const v5 = require('@unlock-protocol/unlock-abi-5')
 const v13 = require('@unlock-protocol/unlock-abi-6')
 const v7 = require('@unlock-protocol/unlock-abi-7')
 /* eslint-enable import/no-extraneous-dependencies */
@@ -27,7 +27,7 @@ const toCompress = {
   v2,
   v3,
   v4,
-  v12,
+  v5,
   v13,
   v7,
 }