+++ unlock-js/src/__tests__/helpers/walletServiceHelper.js
@@ -3,7 +3,7 @@ import * as UnlockV1 from '@unlock-protocol/unlock-abi-1'
 import * as UnlockV2 from '@unlock-protocol/unlock-abi-2'
 import * as UnlockV3 from '@unlock-protocol/unlock-abi-3'
 import * as UnlockV4 from '@unlock-protocol/unlock-abi-4'
-import * as UnlockV12 from '@unlock-protocol/unlock-abi-5'
+import * as UnlockV5 from '@unlock-protocol/unlock-abi-5'
 import * as UnlockV13 from '@unlock-protocol/unlock-abi-6'
 import * as UnlockV7 from '@unlock-protocol/unlock-abi-7'
 import { ethers } from 'ethers'
@@ -56,8 +56,8 @@ export const prepWalletService = async (
       unlockVersion =
         '0x0000000000000000000000000000000000000000000000000000000000000004'
       break
-    case UnlockV12.Unlock:
-    case UnlockV12.PublicLock:
+    case UnlockV5.Unlock:
+    case UnlockV5.PublicLock:
       unlockVersion =
         '0x0000000000000000000000000000000000000000000000000000000000000005'
       break