+++ unlock-js/src/__tests__/v5/isLockManager.test.js
@@ -1,4 +1,4 @@
-import v12 from '../../v12'
+import v5 from '../../v5'
 
 import Web3Service from '../../web3Service'
 import { getTestLockContract } from '../helpers/contracts'
@@ -11,13 +11,13 @@ const owner = '0x90F8bf6A479f320ead074411a4B0e7944Ea8c9C1'
 
 const lockContract = getTestLockContract({
   lockAddress,
-  abi: v12.PublicLock.abi,
+  abi: v5.PublicLock.abi,
   provider,
 })
 
 let web3Service
 
-describe('v12', () => {
+describe('v5', () => {
   describe('isLockManager', () => {
     beforeEach(() => {
       web3Service = new Web3Service({
@@ -25,7 +25,7 @@ describe('v12', () => {
         network: 1984,
       })
 
-      web3Service.lockContractAbiVersion = jest.fn(() => Promise.resolve(v12))
+      web3Service.lockContractAbiVersion = jest.fn(() => Promise.resolve(v5))
       web3Service.getLockContract = jest.fn(() => Promise.resolve(lockContract))
       lockContract.owner = jest.fn(() => {
         return owner